Cleanup rendering code
