+++ packages/reporters/cli/package.json
@@ -41,6 +41,7 @@
     "filesize": "^6.1.0",
     "nullthrows": "^1.1.1",
     "ora": "^5.2.0",
+    "slice-ansi": "^4.0.0",
     "string-width": "^4.2.0",
     "strip-ansi": "^6.0.0",
     "wrap-ansi": "^7.0.0"
