+++ packages/reporters/cli/src/CLIReporter.js
@@ -8,10 +8,12 @@ import {prettifyTime, prettyDiagnostic, throttle} from '@parcel/utils';
 import chalk from 'chalk';
 import {applyDiagnosticFix} from '@parcel/diagnostic';
 
-import {getProgressMessage, getTerminalWidth} from './utils';
+import {getProgressMessage, wrapWithIndent, indentString} from './utils';
 import logLevels from './logLevels';
 import bundleReport from './bundleReport';
 import {
+  init,
+  exit,
   writeOut,
   updateSpinner,
   persistSpinner,
@@ -20,13 +22,14 @@ import {
   persistMessage,
   getLocation,
   moveBy,
-  updateLine,
   scrollTo,
   getPageHeight,
   scrollIfNeeded,
   addFooterLine,
   updateFooterLine,
+  terminalSize,
 } from './render';
+import {initFixes, resetFixes, renderFix, renderFooter} from './fix';
 import * as emoji from './emoji';
 import wrapAnsi from 'wrap-ansi';
 
@@ -35,8 +38,6 @@ const seenWarnings = new Set();
 const seenPhases = new Set();
 
 let isWatching = false;
-let buttons = [];
-let activeButton = -1;
 
 let statusThrottle = throttle((message: string) => {
   updateSpinner(message);
@@ -53,11 +54,8 @@ export async function _report(
     case 'watchStart':
       isWatching = true;
 
-      if (isTTY) {
-        process.stderr.write('\x1b[?1049h'); // use alternate buffer
-        process.stderr.write('\x1b[?1h'); // enable application cursor keys
-        process.stderr.write('\x1b='); // enable application keypad
-      }
+      init();
+      initFixes();
 
       if (options.serveOptions) {
         persistMessage(
@@ -74,11 +72,7 @@ export async function _report(
       break;
 
     case 'watchEnd':
-      if (isTTY) {
-        process.stderr.write('\x1b>'); // disable application keypad
-        process.stderr.write('\x1b[?1l'); // disable application cursor keys
-        process.stderr.write('\x1b[?1049l'); // disable alternate buffer
-      }
+      exit();
       break;
 
     case 'buildStart': {
@@ -90,6 +84,7 @@ export async function _report(
 
       // Clear any previous output
       resetWindow();
+      resetFixes();
 
       break;
     }
@@ -151,6 +146,7 @@ export async function _report(
       }
 
       resetWindow();
+      resetFixes();
 
       persistSpinner('buildProgress', 'error', chalk.red.bold('Build failed.'));
 
@@ -200,14 +196,7 @@ async function writeDiagnostic(
   color: Color,
   isError: boolean = false,
 ) {
-  buttons.length = 0;
-
-  if (isTTY) {
-    // TODO: why doesn't this work during initialization?
-    process.stderr.write('\x1b[?25l'); // hide cursor
-  }
-
-  let columns = getTerminalWidth().columns;
+  let columns = terminalSize.columns;
   let indent = 2;
   for (let diagnostic of diagnostics) {
     let {message, stack, codeframe, hints, fixes, documentation} =
@@ -277,197 +266,10 @@ async function writeDiagnostic(
     }
   }
 
-  if (buttons.length) {
-    addFooterLine('');
-    addFooterLine(' ' + chalk.gray('─'.repeat(columns - 2)) + ' ');
-    addFooterLine('');
-    let allAccepted = buttons.slice(0, -1).filter(b => b.accept);
-    addButton({
-      line: 3,
-      scrollLine: 3,
-      column: 2,
-      text: chalk.underline(
-        `Apply ${allAccepted.length} selected ${
-          allAccepted.length === 1 ? 'fix' : 'fixes'
-        }`,
-      ),
-      disabled: false,
-      footer: true,
-      action: async () => {
-        for (let button of buttons.slice(0, -1)) {
-          await applyDiagnosticFix(button.fix, options.inputFS);
-        }
-      },
-    });
-
-    addFooterLine('');
-  } else if (isError) {
+  let hasFooter = renderFooter(options);
+  if (!hasFooter && isError) {
     writeOut('', isError);
   }
-
-  if (buttons.length) {
-    updateButtons(buttons.length - 1);
-  }
-}
-
-import nullthrows from 'nullthrows';
-import {formatDiff} from '@parcel/codeframe';
-import path from 'path';
-
-async function renderFix(fix, options, {index, length} = {}) {
-  if (fix.type === 'patch') {
-    let code = await options.inputFS.readFile(nullthrows(fix.filePath), 'utf8');
-    let formatted = formatDiff(code, fix.edits, {
-      useColor: true,
-      syntaxHighlighting: true,
-      language:
-        fix.filePath != null ? path.extname(fix.filePath).substr(1) : undefined,
-      terminalWidth: getTerminalWidth().columns,
-    });
-
-    let message =
-      chalk.bold.blue((index == null ? 'Fix: ' : `${index}) `) + fix.message) +
-      '\n\n';
-
-    let location = `${fix.filePath}:${fix.edits[0].range.start.line}:${fix.edits[0].range.start.column}`;
-    message += chalk.gray.underline(location) + '\n';
-
-    if (index == null) {
-      message = emoji.hint + ' ' + message;
-    }
-
-    // result.fixes.push(message + formatted);
-    let fixLine = getLocation();
-    writeOut(
-      wrapWithIndent(
-        message + formatted,
-        index == null ? 5 : 8,
-        index == null ? 2 : 5,
-      ),
-    );
-
-    if (isWatching && isTTY) {
-      writeOut('', true);
-      let buttonIndex = buttons.length;
-      let accept = index == null ? true : index === 1;
-      let button = {
-        line: getLocation(),
-        scrollLine: fixLine,
-        column: index == null ? 5 : 8,
-        text: (accept ? '◉ ' : '◯ ') + chalk.underline('Accept fix'),
-        accept,
-        fix,
-        async action() {
-          if (index == null) {
-            button.accept = !button.accept;
-            button.text =
-              (button.accept ? '◉ ' : '◯ ') + chalk.underline('Accept fix');
-            writeButton(button, true);
-          } else {
-            let baseIndex = buttonIndex - index + 1;
-            for (let i = 0; i < length; i++) {
-              let b = buttons[baseIndex + i];
-              if (i === index - 1) {
-                b.accept = !b.accept;
-              } else {
-                b.accept = false;
-              }
-              b.text = (b.accept ? '◉ ' : '◯ ') + chalk.underline('Accept fix');
-              writeButton(b, i === index - 1);
-            }
-          }
-
-          let allButton = buttons[buttons.length - 1];
-          let allAccepted = buttons.slice(0, -1).filter(b => b.accept);
-          allButton.text = chalk.underline(
-            allAccepted.length === 0
-              ? 'No fixes selected'
-              : `Apply ${allAccepted.length} selected ${
-                  allAccepted.length === 1 ? 'fix' : 'fixes'
-                }`,
-          );
-
-          allButton.disabled = allAccepted.length === 0;
-          writeButton(allButton, false);
-        },
-      };
-      addButton(button);
-    }
-  } else if (fix.type === 'group') {
-    writeOut(
-      wrapWithIndent(
-        chalk.bold.blue(`${emoji.hint} ${fix.options.length} possible fixes:`),
-        5,
-        2,
-      ) + '\n',
-    );
-    for (let [index, option] of fix.options.entries()) {
-      await renderFix(option, options, {
-        index: index + 1,
-        length: fix.options.length,
-      });
-      writeOut('\n');
-    }
-  }
-}
-
-import readline from 'readline';
-
-function addButton(button) {
-  buttons.push(button);
-  writeButton(button, false);
-}
-
-function writeButton(button, focused) {
-  let update = button.footer ? updateFooterLine : updateLine;
-
-  if (button.disabled) {
-    if (focused) {
-      update(
-        button.column,
-        button.line,
-        chalk.bgBlackBright.hex('#fff')(chalk(button.text)),
-      );
-    } else {
-      update(button.column, button.line, chalk.gray(button.text));
-    }
-  } else {
-    if (focused) {
-      update(
-        button.column,
-        button.line,
-        chalk.bgBlue
-          .hex('#fff')
-          .bold(chalk(button.text) + (button.accept != null ? '' : ' ↵')),
-      );
-    } else {
-      update(button.column, button.line, chalk.bold.blue(button.text));
-    }
-  }
-}
-
-function updateButtons(newButton) {
-  if (activeButton >= 0) {
-    writeButton(buttons[activeButton], false);
-  }
-
-  writeButton(buttons[newButton], true);
-  activeButton = newButton;
-}
-
-function wrapWithIndent(string, indent = 0, initialIndent = indent) {
-  let width = getTerminalWidth().columns;
-  return indentString(
-    wrapAnsi(string.trimEnd(), width - indent, {trim: false}),
-    indent,
-    initialIndent,
-  );
-}
-
-function indentString(string, indent = 0, initialIndent = indent) {
-  return (
-    ' '.repeat(initialIndent) + string.replace(/\n/g, '\n' + ' '.repeat(indent))
-  );
 }
 
 export default (new Reporter({
@@ -475,73 +277,3 @@ export default (new Reporter({
     return _report(event, options);
   },
 }): Reporter);
-
-process.stdin.setRawMode(true);
-process.stdin.resume();
-process.stdin.setEncoding('utf8');
-
-process.stdin.on('data', key => {
-  if (buttons.length === 0) return;
-
-  for (let i = 0; i < key.length; ) {
-    if (isKey(key, i, '\x1b[Z')) {
-      // tab backward
-      let b = activeButton - 1;
-      if (b < 0) {
-        b = buttons.length - 1;
-      }
-      if (!buttons[b].footer) {
-        scrollIfNeeded(buttons[b].scrollLine);
-        scrollIfNeeded(buttons[b].line);
-      }
-      updateButtons(b);
-      i += 3;
-    } else if (isKey(key, i, '\t') || isKey(key, i, '\x1b[I')) {
-      // tab forward
-      let b = (activeButton + 1) % buttons.length;
-      if (!buttons[b].footer) {
-        scrollIfNeeded(buttons[b].scrollLine);
-        scrollIfNeeded(buttons[b].line);
-      }
-      updateButtons(b);
-      i += key[i] === '\t' ? 1 : 3;
-    } else if (isKey(key, i, '\r')) {
-      // enter
-      let button = buttons[activeButton];
-      if (!button.disabled) {
-        button.action();
-      }
-      i++;
-    } else if (isKey(key, i, '\x1bOB')) {
-      // down arrow
-      moveBy(1);
-      i += 3;
-    } else if (isKey(key, i, '\x1bOA')) {
-      // up arrow
-      moveBy(-1);
-      i += 3;
-    } else if (isKey(key, i, '\x1b[6~')) {
-      // page down
-      moveBy(getPageHeight());
-      i += 4;
-    } else if (isKey(key, i, '\x1b[5~')) {
-      // page up
-      moveBy(-getPageHeight());
-      i += 4;
-    } else if (isKey(key, i, '\x1bOH')) {
-      // home
-      scrollTo(0);
-      i += 3;
-    } else if (isKey(key, i, '\x1bOF')) {
-      // end
-      scrollTo(Infinity);
-      i += 3;
-    } else {
-      i++;
-    }
-  }
-});
-
-function isKey(s: string, i: number, k: string): boolean {
-  return s.slice(i, i + k.length) === k;
-}
